% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{calc_cumulative_alpha}
\alias{calc_cumulative_alpha}
\title{Calculate the cumulative alpha as the individual alpha are combined}
\usage{
calc_cumulative_alpha(individual_alpha)
}
\arguments{
\item{individual_alpha}{numeric vector of alpha values which will be
superimposed.}
}
\value{
numeric vector of the cumulative alpha as these elements are
        rendered on top of each other
}
\description{
Remember:
}
\details{
\itemize{
\item{alpha is not additive}
\item{transmissivity = 1 - alpha}
\item{transmissivity is multiplicative}
}
}
