% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{calc_individual_alpha}
\alias{calc_individual_alpha}
\title{Calculate the individual alpha levels to recreate the cumulative alpha series}
\usage{
calc_individual_alpha(cumulative_alpha)
}
\arguments{
\item{cumulative_alpha}{numeric vector of desired cumulative alpha as
multiple objects are rendered on top of each other}
}
\value{
numeric vector of the individual alpha values of the objects to be
        superimposed
}
\description{
Remember:
}
\details{
\itemize{
\item{alpha is not additive}
\item{transmissivity = 1 - alpha}
\item{transmissivity is multiplicative}
}
}
