% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-key-point-blur.R
\name{draw_key_point_blur}
\alias{draw_key_point_blur}
\title{Key glyphs for legends}
\usage{
draw_key_point_blur(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
Each geom has an associated function that draws the key when the geom needs
to be displayed in a legend. These functions are called `draw_key_*()`, where
`*` stands for the name of the respective key glyph. The key glyphs can be
customized for individual geoms by providing a geom with the `key_glyph`
argument (see [`layer()`] or examples below.)
}
