% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-point-blur.R
\name{geom_point_blur}
\alias{geom_point_blur}
\title{Draw Blurry Points}
\usage{
geom_point_blur(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  blur_size = 3,
  blur_steps = 20,
  blur_alpha = 1,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, position, ..., na.rm, show.legend, inherit.aes}{See
documentation for \code{ggplot2::geom_point()}}

\item{blur_size}{ow far the blur extends from the edge of the drawn point.
Default: 3}

\item{blur_steps}{Number of point repetitions to create blur. Default: 20}

\item{blur_alpha}{The alpha for the inner-most part of the blur. Note: this
is only approximate as the actual rendering of multiple layers with small alpha
does not always result in the expected alpha.  Values above 1 are allowed
here. Please experiment to achieve the desired results. Default: 1}
}
\description{
Draw Blurry Points
}
\examples{
\dontrun{
ggplot(mtcars) +
geom_point_blur(aes(mpg, wt, blur_size = disp), blur_alpha = 0.3, blur_steps = 2) +
  scale_blur_size_continuous(range = c(1, 15)) +
  theme_bw() +
  labs(title = "Larger blur indicates larger engine displacement")
}
}
