% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-blur-size.R
\name{scale_blur_size_continuous}
\alias{scale_blur_size_continuous}
\alias{scale_blur_size}
\alias{scale_blur_size_discrete}
\alias{scale_blur_size_manual}
\title{Scale blur size}
\usage{
scale_blur_size_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(3, 9),
  trans = "identity",
  guide = FALSE
)

scale_blur_size(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(3, 9),
  trans = "identity",
  guide = FALSE
)

scale_blur_size_discrete(..., range = c(4, 9), guide = FALSE)

scale_blur_size_manual(..., values)
}
\arguments{
\item{name, breaks, labels, limits, range, trans, values, ...}{see \code{?ggplot2::scale_size()}}

\item{guide}{default: FALSE i.e. don't draw guide for blur.}
}
\description{
Scale blur size
}
